bugzilla: generalize statuses to bug states
The Bugzilla plugin has four fields to configure "statuses": either
they are statuses that the BZ expects referenced bugs to have, or they
specify the status to which the bugs would be moved after some action:

```
  ...
  statuses:
  - VALIDATED
  dependent_bug_statuses:
  - VALIDATED
  status_after_validation: "PRE"
  status_after_merge: MODIFIED`
```

This ability is insuficient as soon as "resolutions" come to play:
resolutions are sub-states of BZ closed statuses. Therefore, this
commit generalizes a status into a "state", which is used for both
purposes (bugs being checked against and updated to).

```
  ...
  valid_states:
  - status: VALIDATED
  - status: CLOSED
    resolution: ERRATA
  dependent_bug_states:
  - states: VALIDATED
  state_after_validation:
    status: "PRE"
  state_after_merge:
    status: MODIFIED`
```

The old "status" config fields are kept to avoid breaking existing
instances and converted to states at runtime.