trigger: /retest should force failed jobs to rerun
The interaction between `/retest` and conditionally executed jobs is
currently confusing. At least some users expect `/retest` to trigger all
failed jobs, even those that only run conditionally (`always_run: false`),
which was not true: the trigger criteria still applied.

I think that the expectation is valid though: once the job was for any
reason triggered and contributed a status to a PR, I find the
expectation that `/retest` will retest it reasonable.

This also makes the interaction with retest bot more intuitive. That bot
searches for jobs with failing tests and posts `/retest` on them. We had
saw a PR with a single contitionally executed job (manual trigger only)
failed, and the bot just continued to attempt to `/retest` without any
effect:

https://github.com/openshift/origin/pull/25195